requirements.txt
absl-py==2.3.0
aiosignal==1.3.2
ale-py==0.11.1
annotated-types==0.7.0
argcomplete==3.4.0
attrs==25.3.0
-e git+https://github.com/AVSLab/basilisk@0f6e3c5406c9bee753517e60f9edf3836eae04d7#egg=Basilisk
bokeh==3.7.2
-e git+https://github.com/AVSLab/bsk_rl@f70bd259049f2bcdc40c76eb258736968bf98cf2#egg=bsk_rl
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
clang-format==20.1.0
click==8.1.7
cloudpickle==3.1.1
cmake==3.31.6
colorama==0.4.6
conan==2.15.1
contourpy==1.3.2
coverage==7.8.2
cycler==0.12.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.8.0
dm-tree==0.1.9
etils==1.12.2
execnet==2.1.1
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
grpcio==1.72.1
gymnasium==1.1.1
identify==2.6.12
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.1.0
jax-jumpy==1.0.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.3.2
narwhals==1.41.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ormsgpack==1.7.0
packaging==24.1
pandas==2.2.3
patch-ng==1.18.1
pettingzoo==1.25.0
pillow==11.2.1
pipx==1.6.0
platformdirs==4.2.2
pluggy==1.6.0
pre_commit==4.2.0
protobuf==6.30.2
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pytest==7.3.1
pytest-cov==6.1.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-repeat==0.9.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.35.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.12
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==78.1.0
setuptools-scm==8.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tifffile==2025.6.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.0.7
userpath==1.9.2
virtualenv==20.31.2
wandb==0.20.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xyzservices==2025.4.0
zipp==3.22.0
